Restructured some classes to make code easier to use from the BQ Pushdown implementation.
